# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: DoleSe Wonderland FX Trading Platform API
  description: Comprehensive API for forex trading education platform with course management, student enrollment, payments, and analytics
  version: 1.0.0
  contact:
    name: DoleSe Wonderland FX
    email: support@dolesefx.com
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.dolesefx.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Advanced Trading Strategies
        slug:
          type: string
          example: advanced-trading-strategies
        description:
          type: string
          example: Master advanced forex trading techniques
        category:
          type: string
          example: Trading
        level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: advanced
        price:
          type: number
          format: decimal
          example: 149.99
        currency:
          type: string
          example: USD
        status:
          type: string
          enum: [draft, published, archived]
          example: published
        thumbnailUrl:
          type: string
          nullable: true
        trailerUrl:
          type: string
          nullable: true
        instructorId:
          type: integer
          example: 5
        averageRating:
          type: number
          format: float
          nullable: true
          example: 4.7
        totalStudents:
          type: integer
          example: 250
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Enrollment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 10
        courseId:
          type: integer
          example: 5
        progressPercentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 65
        completed:
          type: boolean
          example: false
        completedAt:
          type: string
          format: date-time
          nullable: true
        certificateUrl:
          type: string
          nullable: true
        enrolledAt:
          type: string
          format: date-time
...
paths:
...
  /courses:
    get:
      tags:
        - Courses
      summary: Get all published courses
      description: Retrieve list of all published courses with filtering and pagination
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: level
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: minPrice
          in: query
          description: Minimum price filter
          schema:
            type: number
        - name: maxPrice
          in: query
          description: Maximum price filter
          schema:
            type: number
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      courses:
                        type: array
                        items:
                          $ref: '#/components/schemas/Course'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 10
                          total:
                            type: integer
                            example: 45
                          pages:
                            type: integer
                            example: 5

    post:
      tags:
        - Courses
      summary: Create new course
      description: Create a new course (instructor/admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - category
                - price
              properties:
                title:
                  type: string
                  example: Advanced Trading Strategies
                description:
                  type: string
                  example: Master advanced forex trading techniques
                category:
                  type: string
                  example: Trading
                level:
                  type: string
                  enum: [beginner, intermediate, advanced]
                  default: beginner
                price:
                  type: number
                  example: 149.99
                currency:
                  type: string
                  default: USD
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      course:
                        $ref: '#/components/schemas/Course'
        '403':
          description: Forbidden - instructor/admin role required
...
  /students/enroll:
    post:
      tags:
        - Students
      summary: Enroll in course
      description: Enroll current user in a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseId
              properties:
                courseId:
                  type: integer
                  example: 5
                paymentId:
                  type: integer
                  example: 123
      responses:
        '201':
          description: Enrollment successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      enrollment:
                        $ref: '#/components/schemas/Enrollment'
        '400':
          description: Already enrolled or invalid course
...